+++ docs/sync.js
@@ -252,6 +252,12 @@ class Sync {
      * }
      */
     static initiateClientReset(path) {}
+
+    /**
+     * Returns `true` if Realm still has a reference to any sync sessions regardless of their state.
+     * If `false` is returned it means that no sessions currently exist.
+     */
+    static _hasExistingSessions() {}
 }
 
 /**